+++ SelectionProcessor/include/Variables.h
@@ -13,12 +13,13 @@
 #include <limits>
 #include <vector>
 
+#include "TInterpreter.h"
 #include "TTree.h"
 
 namespace variables_namespace
 {
     typedef std::vector<int> IntVector;
-    typedef std::vector<float> FloatVector;
+    typedef std::vector<double> DoubleVector;
 }
 
 using namespace variables_namespace;
@@ -44,7 +45,7 @@ class Variables
         void SetBranches(TTree *pTTree);
 
         /**
-         *  @brief Print to screen all bool, int, float and selected vector member variables
+         *  @brief Print to screen all bool, int, double and selected vector member variables
          */
         void Print();
 
@@ -160,307 +161,307 @@ class Variables
          *
          *  @param transverseMomentum Transverse momentum of reconstructed system
          */
-        void SetTransverseMomentum(float transverseMomentum);
+        void SetTransverseMomentum(double transverseMomentum);
 
         /**
          *  @brief Get m_TransverseMomentum
          */
-        float GetTransverseMomentum() const;
+        double GetTransverseMomentum() const;
 
         /**
          *  @brief Set m_MCTransverseMomentum
          *
          *  @param mcTransverseMomentum Transverse momentum of quark system
          */
-        void SetMCTransverseMomentum(float mcTransverseMomentum);
+        void SetMCTransverseMomentum(double mcTransverseMomentum);
 
         /**
          *  @brief Get m_MCTransverseMomentum
          */
-        float GetMCTransverseMomentum() const;
+        double GetMCTransverseMomentum() const;
 
         /**
          *  @brief Set m_TransverseEnergy
          *
          *  @param transverseEnergy Transverse energy of reconstructed system
          */
-        void SetTransverseEnergy(float transverseEnergy);
+        void SetTransverseEnergy(double transverseEnergy);
 
         /**
          *  @brief Get m_TransverseEnergy
          */
-        float GetTransverseEnergy() const;
+        double GetTransverseEnergy() const;
 
         /**
          *  @brief Set m_MCTransverseEnergy
          *
          *  @param mcTransverseEnergy Transverse energy of quark system
          */
-        void SetMCTransverseEnergy(float mcTransverseEnergy);
+        void SetMCTransverseEnergy(double mcTransverseEnergy);
 
         /**
          *  @brief Get m_MCTransverseEnergy
          */
-        float GetMCTransverseEnergy() const;
+        double GetMCTransverseEnergy() const;
 
         /**
          *  @brief Set m_CosThetaMissing
          *
          *  @param cosThetaMissing Cosine theta of missing energy
          */
-        void SetCosThetaMissing(float cosThetaMissing);
+        void SetCosThetaMissing(double cosThetaMissing);
 
         /**
          *  @brief Get m_CosThetaMissing
          */
-        float GetCosThetaMissing() const;
+        double GetCosThetaMissing() const;
 
         /**
          *  @brief Set m_MCCosThetaMissing
          *
          *  @param mcCosThetaMissing Cosine theta of neutrino energy
          */
-        void SetMCCosThetaMissing(float mcCosThetaMissing);
+        void SetMCCosThetaMissing(double mcCosThetaMissing);
 
         /**
          *  @brief Get m_MCCosThetaMissing
          */
-        float GetMCCosThetaMissing() const;
+        double GetMCCosThetaMissing() const;
 
         /**
          *  @brief Set m_CosThetaMostEnergeticTrack 
          *
          *  @param cosThetaMostEnergeticTrack Cosine theta of highest energy track in system
          */
-        void SetCosThetaMostEnergeticTrack(float cosThetaMostEnergeticTrack);
+        void SetCosThetaMostEnergeticTrack(double cosThetaMostEnergeticTrack);
 
         /** 
          *  @brief Get m_CosThetaMostEnergeticTrack
          */
-        float GetCosThetaMostEnergeticTrack() const;
+        double GetCosThetaMostEnergeticTrack() const;
 
         /**
          *  @brief Set m_RecoilMass
          *
          *  @param recoilMass Recoil mass of reconstructed system
          */
-        void SetRecoilMass(float recoilMass);
+        void SetRecoilMass(double recoilMass);
 
         /**
          *  @brief Get m_RecoilMass
          */
-        float GetRecoilMass() const;
+        double GetRecoilMass() const;
 
         /**
          *  @brief Set m_MCRecoilMass
          *
          *  @param mcRecoilMass Recoil mass of quark system
          */
-        void SetMCRecoilMass(float mcRecoilMass);
+        void SetMCRecoilMass(double mcRecoilMass);
 
         /**
          *  @brief Get m_MCRecoilMass
          */
-        float GetMCRecoilMass() const;
+        double GetMCRecoilMass() const;
 
         /**
          *  @brief Set m_EnergyAroundMostEnergeticPfo
          *
          *  @param energyAroundMostEnergeticPfo Energy in cone around most energetic pfo
          */
-        void SetEnergyAroundMostEnergeticPfo(float energyAroundMostEnergeticPfo);
+        void SetEnergyAroundMostEnergeticPfo(double energyAroundMostEnergeticPfo);
 
         /**
          *  @brief Get m_EnergyAroundMostEnergeticPfo
          */
-        float GetEnergyAroundMostEnergeticPfo() const;
+        double GetEnergyAroundMostEnergeticPfo() const;
 
         /**
          *  @brief Set m_y34
          *
          *  @param y34 Jet grouping parameter
          */
-        void SetY34(float y34);
+        void SetY34(double y34);
 
         /**
          *  @brief Get m_y34
          */
-        float GetY34() const;
+        double GetY34() const;
 
         /**
          *  @brief Set m_InvariantMassSystem
          *
          *  @param invariantMassSystem Invariant mass of the jet system
          */
-        void SetInvariantMassSystem(float invariantMassSystem);
+        void SetInvariantMassSystem(double invariantMassSystem);
 
         /**
          *  @brief Get m_InvariantMassSystem
          */
-        float GetInvariantMassSystem() const;
+        double GetInvariantMassSystem() const;
 
         /**
          *  @brief Set m_MCInvariantMassSystem
          *
          *  @param mcInvariantMassSystem Invariant mass of the quark system
          */
-        void SetMCInvariantMassSystem(float mcInvariantMassSystem);
+        void SetMCInvariantMassSystem(double mcInvariantMassSystem);
 
         /**
          *  @brief Get m_MCInvariantMassSystem
          */
-        float GetMCInvariantMassSystem() const;
+        double GetMCInvariantMassSystem() const;
 
         /**
          *  @brief Set m_CosThetaStarWBosons
          *
          *  @param cosThetaStarWBosons Cosine of the polar angle of the W bosons in the reference frame of the WW boson pairs
          */
-        void SetCosThetaStarWBosons(float cosThetaStarWBosons);
+        void SetCosThetaStarWBosons(double cosThetaStarWBosons);
 
         /**
          *  @brief Get m_CosThetaStarWBosons
          */
-        float GetCosThetaStarWBosons() const;
+        double GetCosThetaStarWBosons() const;
 
         /**
          *  @brief Set m_MCCosThetaStarWBosons
          *
          *  @param mcCosThetaStarWBosons Cosine of the polar angle of the W bosons in the reference frame of the WW boson pairs
          */
-        void SetMCCosThetaStarWBosons(float mcCosThetaStarWBosons);
+        void SetMCCosThetaStarWBosons(double mcCosThetaStarWBosons);
 
         /**
          *  @brief Get m_MCCosThetaStarWBosons
          */
-        float GetMCCosThetaStarWBosons() const;
+        double GetMCCosThetaStarWBosons() const;
 
         /**
          *  @brief Set m_CosThetaStarZBosons
          *
          *  @param cosThetaStarZBosons Cosine of the polar angle of the Z bosons in the reference frame of the ZZ boson pairs
          */
-        void SetCosThetaStarZBosons(float cosThetaStarZBosons);
+        void SetCosThetaStarZBosons(double cosThetaStarZBosons);
 
         /**
          *  @brief Get m_CosThetaStarZBosons
          */
-        float GetCosThetaStarZBosons() const;
+        double GetCosThetaStarZBosons() const;
 
         /**
          *  @brief Set m_MCCosThetaStarZBosons
          *
          *  @param mcCosThetaStarZBosons Cosine of the polar angle of the Z bosons in the reference frame of the ZZ boson pairs
          */
-        void SetMCCosThetaStarZBosons(float mcCosThetaStarZBosons);
+        void SetMCCosThetaStarZBosons(double mcCosThetaStarZBosons);
 
         /**
          *  @brief Get m_MCCosThetaStarZBosons
          */
-        float GetMCCosThetaStarZBosons() const;
+        double GetMCCosThetaStarZBosons() const;
 
-// FloatVectors
+// DoubleVectors
 
         /**
          *  @brief Set m_InvMassWVectors
          *
          *  @param invMassWVectors Invariant mass of first and second pair of W paired jets
          */
-        void SetInvMassWVectors(FloatVector invMassWVectors);
+        void SetInvMassWVectors(DoubleVector invMassWVectors);
 
         /**
          *  @brief Get m_InvMassWVectors
          */
-        FloatVector GetInvMassWVectors() const;
+        DoubleVector GetInvMassWVectors() const;
 
         /**
          *  @brief Set m_InvMassZVectors 
          *
          *  @param invMassZVectors Invariant mass of first and second pair of Z paired jets
          */
-        void SetInvMassZVectors(FloatVector invMassZVectors);
+        void SetInvMassZVectors(DoubleVector invMassZVectors);
 
         /**
          *  @brief Get m_InvMassZVectors
          */
-        FloatVector GetInvMassZVectors() const;
+        DoubleVector GetInvMassZVectors() const;
 
         /**
          *  @brief Set m_MCInvMassWVectors
          *
          *  @param invMassWVectors Invariant mass of first and second pair of W paired quarks
          */
-        void SetMCInvMassWVectors(FloatVector mcInvMassWVectors);
+        void SetMCInvMassWVectors(DoubleVector mcInvMassWVectors);
 
         /**
          *  @brief Get m_MCInvMassWVectors
          */
-        FloatVector GetMCInvMassWVectors() const;
+        DoubleVector GetMCInvMassWVectors() const;
 
         /**
          *  @brief Set m_MCInvMassZVectors
          *
          *  @param mcInvMassZVectors Invariant mass of first and second pair of Z paired quarks
          */
-        void SetMCInvMassZVectors(FloatVector mcInvMassZVectors);
+        void SetMCInvMassZVectors(DoubleVector mcInvMassZVectors);
 
         /**
          *  @brief Get m_MCInvMassZVectors
          */
-        FloatVector GetMCInvMassZVectors() const;
+        DoubleVector GetMCInvMassZVectors() const;
 
         /**
          *  @brief Set m_EnergyJets
          *
          *  @param energyJets Energy of reconstructed jets
          */
-        void SetEnergyJets(FloatVector energyJets);
+        void SetEnergyJets(DoubleVector energyJets);
 
         /**
          *  @brief Get m_EnergyJets
          */
-        FloatVector GetEnergyJets() const;
+        DoubleVector GetEnergyJets() const;
 
         /**
          *  @brief Get lowest number in m_EnergyJets
          */
-        float GetLowestEnergyJet() const;
+        double GetLowestEnergyJet() const;
 
         /**
          *  @brief Set m_CosThetaStarWJets
          *
          *  @param cosThetaStarWJets Cosine of the polar angle of the q jets in the reference frome of the W bosons
          */
-        void SetCosThetaStarWJets(FloatVector cosThetaStarWJets);
+        void SetCosThetaStarWJets(DoubleVector cosThetaStarWJets);
 
         /**
          *  @brief Get m_CosThetaStarWJets
          */
-        FloatVector GetCosThetaStarWJets() const;
+        DoubleVector GetCosThetaStarWJets() const;
 
         /**
          *  @brief Set m_CosThetaStarZJets
          *
          *  @param cosThetaStarZJets Cosine of the polar angle of the q jets in the reference frome of the Z bosons
          */
-        void SetCosThetaStarZJets(FloatVector cosThetaStarZJets);
+        void SetCosThetaStarZJets(DoubleVector cosThetaStarZJets);
 
         /**
          *  @brief Get m_CosThetaStarZJets
          */
-        FloatVector GetCosThetaStarZJets() const;
+        DoubleVector GetCosThetaStarZJets() const;
 
         /**
          *  @brief Set
          *
          *  @param
          */
-//        void Set(FloatVector );
+//        void Set(DoubleVector );
 
         /**
          *  @brief Get
          */
-//        FloatVector Get();
+//        DoubleVector Get();
 
     private:
         bool            m_AppropriateEvent;                 ///< Does event pass a series of cuts suggesting it has a vvqqqq final state? - JetAnalysis
@@ -472,31 +473,31 @@ class Variables
         IntVector       m_NParticlesJets;                   ///< Number of particles in reconstructed jet - JetAnalysis
         IntVector       m_NChargedParticlesJets;            ///< Number of charged particles in reconstructed jet - JetAnalysis
 
-        float           m_TransverseMomentum;               ///< Transverse momentum of reconstructed system - JetAnalysis
-        float           m_MCTransverseMomentum;             ///< Transverse momentum of quark system - MCAnalysis
-        float           m_TransverseEnergy;                 ///< Transverse energy of reconstructed system - JetAnalysis
-        float           m_MCTransverseEnergy;               ///< Transverse energy of quark system - MCAnalysis
-        float           m_CosThetaMissing;                  ///< Cosine theta of missing energy - JetAnalysis
-        float           m_MCCosThetaMissing;                ///< Cosine theta of neutrino energy - MCAnalysis
-        float           m_CosThetaMostEnergeticTrack;       ///< Cosine theta of highest energy track in system - JetAnalysis
-        float           m_RecoilMass;                       ///< Recoil mass of reconstructed system - JetAnalysis
-        float           m_MCRecoilMass;                     ///< Recoil mass of quark system - MCAnalysis
-        float           m_EnergyAroundMostEnergeticPfo;     ///< Energy in cone around most energetic pfo - JetAnalysis
-        float           m_y34;                              ///< Jet grouping parameter
-        float           m_InvariantMassSystem;              ///< Invariant mass of the jet system - JetAnalysis
-        float           m_MCInvariantMassSystem;            ///< Invariant mass of the quark system - MCAnalysis
-        float           m_CosThetaStarWBosons;              ///< Cosine of the polar angle of the W bosons in the reference frame of the WW boson pairs - JetAnalysis
-        float           m_MCCosThetaStarWBosons;            ///< Cosine of the polar angle of the W bosons in the reference frame of the WW boson pairs - MCAnalysis
-        float           m_CosThetaStarZBosons;              ///< Cosine of the polar angle of the Z bosons in the reference frame of the ZZ boson pairs - JetAnalysis
-        float           m_MCCosThetaStarZBosons;            ///< Cosine of the polar angle of the Z bosons in the reference frame of the ZZ boson pairs - MCAnalysis
+        double           m_TransverseMomentum;               ///< Transverse momentum of reconstructed system - JetAnalysis
+        double           m_MCTransverseMomentum;             ///< Transverse momentum of quark system - MCAnalysis
+        double           m_TransverseEnergy;                 ///< Transverse energy of reconstructed system - JetAnalysis
+        double           m_MCTransverseEnergy;               ///< Transverse energy of quark system - MCAnalysis
+        double           m_CosThetaMissing;                  ///< Cosine theta of missing energy - JetAnalysis
+        double           m_MCCosThetaMissing;                ///< Cosine theta of neutrino energy - MCAnalysis
+        double           m_CosThetaMostEnergeticTrack;       ///< Cosine theta of highest energy track in system - JetAnalysis
+        double           m_RecoilMass;                       ///< Recoil mass of reconstructed system - JetAnalysis
+        double           m_MCRecoilMass;                     ///< Recoil mass of quark system - MCAnalysis
+        double           m_EnergyAroundMostEnergeticPfo;     ///< Energy in cone around most energetic pfo - JetAnalysis
+        double           m_y34;                              ///< Jet grouping parameter
+        double           m_InvariantMassSystem;              ///< Invariant mass of the jet system - JetAnalysis
+        double           m_MCInvariantMassSystem;            ///< Invariant mass of the quark system - MCAnalysis
+        double           m_CosThetaStarWBosons;              ///< Cosine of the polar angle of the W bosons in the reference frame of the WW boson pairs - JetAnalysis
+        double           m_MCCosThetaStarWBosons;            ///< Cosine of the polar angle of the W bosons in the reference frame of the WW boson pairs - MCAnalysis
+        double           m_CosThetaStarZBosons;              ///< Cosine of the polar angle of the Z bosons in the reference frame of the ZZ boson pairs - JetAnalysis
+        double           m_MCCosThetaStarZBosons;            ///< Cosine of the polar angle of the Z bosons in the reference frame of the ZZ boson pairs - MCAnalysis
 
-        FloatVector     m_InvMassWVectors;                  ///< Invariant mass of first and second pair of W paired jets - JetAnalysis
-        FloatVector     m_MCInvMassWVectors;                ///< Invariant mass of first and second pair of W paired quarks - MCAnalysis
-        FloatVector     m_InvMassZVectors;                  ///< Invariant mass of first and second pair of Z paired jets - JetAnalysis
-        FloatVector     m_MCInvMassZVectors;                ///< Invariant mass of first and second pair of Z paired quarks - MCAnalysis
-        FloatVector     m_EnergyJets;                       ///< Energy of reconstructed jets  - JetAnalysis
-        FloatVector     m_CosThetaStarWJets;                ///< Cosine of the polar angle of the q jets in the reference frome of the W bosons - JetAnalysis
-        FloatVector     m_CosThetaStarZJets;                ///< Cosine of the polar angle of the q jets in the reference frome of the Z bosons - JetAnalysis
+        DoubleVector     m_InvMassWVectors;                  ///< Invariant mass of first and second pair of W paired jets - JetAnalysis
+        DoubleVector     m_MCInvMassWVectors;                ///< Invariant mass of first and second pair of W paired quarks - MCAnalysis
+        DoubleVector     m_InvMassZVectors;                  ///< Invariant mass of first and second pair of Z paired jets - JetAnalysis
+        DoubleVector     m_MCInvMassZVectors;                ///< Invariant mass of first and second pair of Z paired quarks - MCAnalysis
+        DoubleVector     m_EnergyJets;                       ///< Energy of reconstructed jets  - JetAnalysis
+        DoubleVector     m_CosThetaStarWJets;                ///< Cosine of the polar angle of the q jets in the reference frome of the W bosons - JetAnalysis
+        DoubleVector     m_CosThetaStarZJets;                ///< Cosine of the polar angle of the q jets in the reference frome of the Z bosons - JetAnalysis
 };
 
 #endif // #ifndef VARIABLES_H
